Develop

+++ uber-clone/src/main/java/com/backend/uberclone/model/Notification.java
@@ -17,17 +17,22 @@ public class Notification {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
+
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name="recipient_id")
     private User recipient;
+
     @Column
     private LocalDateTime date;
+
     @Column
     private String text;
+
     @Column
     @Enumerated(EnumType.STRING)
     private NotificationType type;
+
     @Column
     private boolean read;
